[Refact] LoginVC를 LoginView로 분류
